% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hello}
\usage{
hello(name = "world")
}
\arguments{
\item{name}{A string. The name to greet.}
}
\value{
A character string of the form \code{"Hello, {name}!"}.
}
\description{
A simple greeting function.
}
\examples{
hello("world")
}
